Finish implementation of zippers
Now this adds goFirst and goSecond, completing the implementation of
Sec. 3.2 of Huet's "Zipper" paper.

However, both functions need to return an existential (which I encode
through CPS), which is slightly annoying. This time however there is no
easy fix, since the needed type is just not known (until we look). The
complicated fix would be to index the tree by a tree of types of the
various members, encoding essentially the "typing derivation" for the
correctness of the tree (in the metalanguage). However, that does not
look trivial.